refactor: rename textfield to input
