+++ dashboard/app/components/settings/Section.module.css
@@ -42,7 +42,7 @@
 
 	padding: 8px;
 	flex: 1;
-	gap: 16px;
+	gap: 4px;
 }
 
 .divider {
