+++ dashboard/app/routes/settings.websites.tsx
@@ -15,7 +15,7 @@ import * as v from 'valibot';
 import { userGet } from '@/api/user';
 import { websiteDelete, websiteList } from '@/api/websites';
 import { ModalChild, ModalWrapper } from '@/components/Modal';
-import { TextInput } from '@/components/TextField';
+import { TextInput } from '@/components/Input';
 import { Group } from '@/components/layout/Flex';
 import {
 	SectionDanger,
